requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Reuploaded==1.2.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.10.0
PyPDF2==2.11.1
pyenchant==3.2.2
pdfplumber==0.7.5
python-dotenv==0.21.0
fuzzywuzzy==0.18.0
docx2python==2.0.4
Flask-Cors==3.0.10
python-Levenshtein==0.20.8
rapidfuzz==2.13.2
python-docx==0.8.11
numpy==1.24.3
python-magic==0.4.27